requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.5
pydantic==2.8.2
python-dotenv==1.0.1
sentence-transformers==3.0.1
chromadb==0.5.5
rank-bm25==0.2.2
nltk==3.9.1
pypdf==4.3.1
pdfplumber==0.11.4
pillow==10.4.0
pytesseract==0.3.13
pdf2image==1.17.0
rapidfuzz==3.9.6
httpx==0.27.2
openai==1.42.0
ollama==0.3.3
